--- a/persistence/soft-index/pom.xml
+++ b/persistence/soft-index/pom.xml
@@ -27,6 +27,16 @@
          <artifactId>mockito-core</artifactId>
          <scope>test</scope>
       </dependency>
+      <dependency>
+         <groupId>org.openjdk.jmh</groupId>
+         <artifactId>jmh-core</artifactId>
+         <scope>test</scope>
+      </dependency>
+      <dependency>
+         <groupId>org.openjdk.jmh</groupId>
+         <artifactId>jmh-generator-annprocess</artifactId>
+         <scope>test</scope>
+      </dependency>
    </dependencies>
 
    <build>