https://issues.redhat.com/browse/ISPN-12481