@franz1981 2020-11-09
Some preliminary numbers

master
```
Benchmark                                  (consumerDelay)  (producerDelay)   Mode  Cnt       Score      Error  Units
SyncProcessingQueueBenchmark.pushAndWait2                0                0  thrpt   16  195281.813 ± 2161.177  ops/s
SyncProcessingQueueBenchmark.pushAndWait2                0               10  thrpt   16  197914.959 ± 2506.130  ops/s
SyncProcessingQueueBenchmark.pushAndWait2                0              100  thrpt   16  187467.698 ± 1885.793  ops/s
SyncProcessingQueueBenchmark.pushAndWait2               10                0  thrpt   16  197724.733 ± 3139.547  ops/s
SyncProcessingQueueBenchmark.pushAndWait2               10               10  thrpt   16  197446.585 ± 3106.995  ops/s
SyncProcessingQueueBenchmark.pushAndWait2               10              100  thrpt   16  189787.057 ± 3239.134  ops/s
SyncProcessingQueueBenchmark.pushAndWait2              100                0  thrpt   16  193570.843 ± 2456.758  ops/s
SyncProcessingQueueBenchmark.pushAndWait2              100               10  thrpt   16  189653.444 ± 2551.090  ops/s
SyncProcessingQueueBenchmark.pushAndWait2              100              100  thrpt   16  183585.022 ± 4243.454  ops/s
```
This pr:
```
Benchmark                                  (consumerDelay)  (producerDelay)   Mode  Cnt        Score        Error  Units
SyncProcessingQueueBenchmark.pushAndWait2                0                0  thrpt   16  1995885.544 ±  55103.406  ops/s
SyncProcessingQueueBenchmark.pushAndWait2                0               10  thrpt   16  2104246.160 ±  77654.840  ops/s
SyncProcessingQueueBenchmark.pushAndWait2                0              100  thrpt   16  2041778.864 ± 190395.636  ops/s
SyncProcessingQueueBenchmark.pushAndWait2               10                0  thrpt   16  2046269.439 ±  60839.916  ops/s
SyncProcessingQueueBenchmark.pushAndWait2               10               10  thrpt   16  1969678.808 ±  39858.488  ops/s
SyncProcessingQueueBenchmark.pushAndWait2               10              100  thrpt   16  1945230.866 ± 228512.231  ops/s
SyncProcessingQueueBenchmark.pushAndWait2              100                0  thrpt   16  1397323.381 ±  19609.587  ops/s
SyncProcessingQueueBenchmark.pushAndWait2              100               10  thrpt   16  1372057.392 ±  27760.970  ops/s
SyncProcessingQueueBenchmark.pushAndWait2              100              100  thrpt   16  1060660.459 ± 241542.295  ops/s
```
Would be interesting to see how this queue behave with a proper super fast pmem disk instead, but I don't have any ;)

The main difference with the original queue is that this one is a bit more memory hungry (but we can blame RxJava mpsc linked q for that, https://github.com/JCTools/JCTools has few gc-free alternatives nudge nudge), but I'm sure this could be improve if necessary.