@ryanemerson 2020-11-26
> Would be interesting to see how this queue behave with a proper super fast pmem disk instead, but I don't have any ;)

@jhalliday Might be able to help!