@franz1981 2020-11-27
Thanks @ryanemerson I have already spoken with @jhalliday about it but I don't know if we would have any chance to try this on a real hw, @jhalliday wdyt? It's feasible to have a round of bench with this in?