@danberindei 2020-12-01
@franz1981 I still haven't managed to delve into this yet, but I have a question.

We also use a `SynchronousQueue` for executors, e.g. in JGroups' [TP.init()](https://github.com/belaban/JGroups/blob/master/src/org/jgroups/protocols/TP.java#L983) and for our expiration executor (the only one that still has 0 queue length). Should we look into changing all of them to a better queue impl?