@franz1981 2020-12-01
@danberindei consider that this queue isn't a full fat blocking queue as synchronous queue nor allow (yet) to fire and and forget tasks as a thread pool impl need.
In general I would say "yes" about having better queue impls (disclaimer: I am one of the devs of https://github.com/JCTools/JCTools) especially if used on AOT given that bad reputation of the jdk ones related to garbage produced.
In term of pure performance, it means of the underlying tasks is expensive in term of time or not, because it means that enqueue/dequeue time start to became a factor.