@franz1981 2020-12-18
Thanks @wburns !
@jhalliday 
Let me know if you got any chance to test it with both a slow and super fast storage so I can provide some sane default (or adaptive, even better) for spin/yield configurations.
On user code instead I believe batching writes before fdatasync/fsynx is a must-have with FileChannels (while pmem won't benefit from it) and should give a nice boost amortizing both syscall pwrite/write and sync cost.